Split staging tests into focused modules
